+++ src/packages/bem/__tests__/dummy-components/index.js
@@ -1,4 +1,4 @@
 export { default as Button } from './Button';
 export { default as ButtonStateless } from './ButtonStateless';
-export { default as Avatar } from './Avatar';
-export { default as AvatarStateless } from './AvatarStateless';
+export { default as List } from './List';
+export { default as Unstyled } from './Unstyled';