+++ stories/index.js
@@ -31,4 +31,5 @@ import './RadioButton';
 import './Teaser';
 // Organisms
 import './Autosuggest';
+import './LocationAutocomplete';
 import './LocationSelector';